back_Sistema_Venta_IA_Rec.Voz/requirements.txt
bcrypt==4.2.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
django>=3.2,<4.0
distlib==0.3.9
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.0.2
fonttools==4.45.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mysql-connector-python==8.3.0
numpy==1.26.2
openpyxl==3.1.5
packaging==23.2
Pillow==10.1.0
platformdirs==4.3.7
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
ttkthemes==3.2.2
virtualenv==20.30.0
Werkzeug==3.0.1


python-dotenv
django-cors-headers
django-environ
django-extensions
psycopg2-binary
djangorestframework==3.14.0
PyJWT==2.8.0

pandas>=2.0
scikit-learn>=1.3
nltk>=3.8